refactor(auth): Consolidate test helpers into mod tests
Move TEST_PASSWORD and mock_config from module scope
into the single #[cfg(test)] mod tests block. Child
modules access them via super::super::tests::{...}
using pub(super) visibility.

Also simplify Python cross-verify CI setup: replace
git clone + pip install -r requirements.txt with just
pip install keystone. Pin actions/setup-python to
commit SHA v5.6.0. Remove KEYSTONE_PYTHON_CHECKOUT
env var; tests now probe `import keystone` at runtime
and skip silently when keystone is not installed.

Note: This commit was done with the help of AI.
