+++ .github/workflows/ci.yml
@@ -58,18 +58,14 @@ jobs:
         uses: swatinem/rust-cache@98c8021b550208e191a6a3145459bfc9fb29c4c0 # v2.8.0
 
       - name: Set up Python for cross-verification tests
-        uses: actions/setup-python@v5
+        uses: actions/setup-python@a26af69be951a213d495a4c3e4e4022e16d87065 # v5.6.0
         with:
           python-version: '3.x'
 
-      - name: Clone Python Keystone and install dependencies
-        run: |
-          git clone --depth=1 https://github.com/openstack/keystone.git /tmp/keystone-python
-          pip install -r /tmp/keystone-python/requirements.txt
+      - name: Install Python Keystone for cross-verification tests
+        run: pip install keystone
 
       - name: Run tests (unit)
-        env:
-          KEYSTONE_PYTHON_CHECKOUT: /tmp/keystone-python
         run: cargo nextest run
 
       - name: Run tests (sqlite)
